pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "calratio_training_data"
version = "0.1.0"
description = "A package for fetching calratio training data"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "awkward",
    "vector",
    "pyarrow",
    "uproot",
    "func_adl_servicex_xaodr25",
    "ServiceX-Local",
    "servicex-analysis-utils",
    "jinja2",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["black", "typer", "pytest", "pytest-mock"]
notebook = ["jupyterlab", "ipywidgets", "hist", "mplhep", "scipy", "pandas"]
local = ["func_adl_xAOD"]

[project.scripts]
calratio_training_data = "calratio_training_data.fetch:run_from_command"

[tool.hatch.build.targets.sdist]
include = ["README.md", "/calratio_training_data"]

[tool.hatch.build.targets.wheel]
include = ["README.md", "/calratio_training_data"]
